# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Wedge Ethiopia ERP API
  version: 0.1.0
  description: Country-neutral ERP surface for fiscal invoice orchestration. Authority-specific fields remain behind the fiscal adapter.
servers:
  - url: https://api.example.et/api/v1
security:
  - ApiKey: []
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    InvoiceLineInput:
      type: object
      required: [description, quantity_milli, unit_price_minor, tax_rate_bps]
      properties:
        sku: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        quantity_milli: { type: integer, minimum: 1, description: Quantity scaled by 1000 }
        unit_price_minor: { type: integer, minimum: 0, description: Minor currency units }
        tax_code: { type: string, maxLength: 80 }
        tax_rate_bps: { type: integer, minimum: 0, description: Tax rate in basis points; authoritative values remain regulator/customer configuration }
    InvoiceCreate:
      type: object
      required: [legal_entity_id, source_ref, lines]
      properties:
        legal_entity_id: { type: string, format: uuid }
        branch_id: { type: string, format: uuid }
        counterparty_id: { type: string, format: uuid }
        source_ref: { type: string, maxLength: 160 }
        currency: { type: string, minLength: 3, maxLength: 3, default: ETB }
        lines:
          type: array
          minItems: 1
          maxItems: 500
          items: { $ref: '#/components/schemas/InvoiceLineInput' }
paths:
...
  /invoices:
    get:
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - in: query
          name: updated_since
          description: ISO-8601 timestamp for incremental ERP synchronization.
          schema: { type: string, format: date-time }
      responses:
        '200': { description: Tenant-scoped invoice collection }
    post:
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvoiceCreate' }
      responses:
        '201': { description: Validated invoice created }
        '400': { description: Invalid payload }
        '401': { description: Invalid API key }
        '403': { description: Missing scope }
        '409': { description: Duplicate source reference or conflicting idempotency key }
...